Account mgmt pundit

+++ app/controllers/highlights_controller.rb
@@ -1,10 +1,12 @@
 class HighlightsController < ApplicationController
   def new
     @scenario_result = ScenarioResult.find_by_hashid(params[:video])
+    authorize @scenario_result, :new_highlight?
   end
 
   def create
     scenario_result = ScenarioResult.find_by_hashid(params[:scenario_highlight][:scenario_result_hashid])
+    authorize scenario_result, :create_highlight?
     scenario = scenario_result.scenario
     @scenario_highlight = service.create(highlight_params, scenario, scenario_result)
     redirect_to highlight_path(@scenario_highlight)
@@ -12,11 +14,13 @@ class HighlightsController < ApplicationController
 
   def edit
     @scenario_highlight = ScenarioHighlight.find_by_hashid!(params[:id])
+    authorize @scenario_highlight
     @scenario_result = @scenario_highlight.scenario_result
   end
 
   def update
     @scenario_highlight = ScenarioHighlight.find_by_hashid!(params[:id])
+    authorize @scenario_highlight
     service.update(@scenario_highlight, highlight_params)
     redirect_to highlight_path(@scenario_highlight)
   end
@@ -29,6 +33,7 @@ class HighlightsController < ApplicationController
 
   def destroy
     @scenario_highlight = ScenarioHighlight.find_by_hashid!(params[:id])
+    authorize @scenario_highlight
     service.destroy(@scenario_highlight)
   end
